Build object model from parse tree
Use alternative labels to make translation easier. Shortens the grammar,
too!

Still having trouble with ANTLR (or the plugin) generating files in
`src/main/gen` (sometimes!) and IntelliJ complaining that the tokens file
isn't available to the parser (which makes me wonder if I actually need to
compile twice... what order are the lexer and parser compiled in?)